hl7_fhir_mapping_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hl7-fhir-mapping-agent"
version = "0.1.0"
description = "AI agent for transforming HL7v2/CCDA to FHIR R4"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Healthcare AI Team" }
]
keywords = ["hl7", "fhir", "healthcare", "ai", "langchain", "langgraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # LLM & Agent Framework
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # HL7 Parsing
    "hl7apy>=1.3.4",
    
    # FHIR
    "fhir.resources>=7.1.0",
    
    # XML Parsing (for CCDA)
    "lxml>=5.0.0",
    
    # Vector Store
    "chromadb>=0.5.0",
    
    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]
demo = [
    "streamlit>=1.40.0",
]

[project.scripts]
hl7-fhir-agent = "src.api.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
